+++ Controller.php
@@ -31,9 +31,12 @@ class Controller extends \Piwik\Plugin\ControllerAdmin
         $view = new View('@McpServer/connect');
         $this->setBasicVariablesView($view);
 
+        $hasSuperUserAccess = Access::getInstance()->hasSuperUserAccess();
+
         $view->assign('isMcpEnabled', $this->systemSettings->isMcpEnabled());
         $view->assign('isOAuth2Enabled', $this->systemSettings->isOAuth2PluginEnabled());
-        $view->assign('canAccessMcpSettings', Access::getInstance()->hasSuperUserAccess());
+        $view->assign('canAccessMcpSettings', $hasSuperUserAccess);
+        $view->assign('canManageOAuth2Clients', $hasSuperUserAccess);
         $view->assign('mcpApiEndpoint', $this->getMcpApiEndpointUrl());
         $view->assign('mcpSettingsUrl', $this->getMcpSettingsUrl());
         $view->assign('oauth2ClientManagementUrl', $this->getOAuth2ClientManagementUrl());